# specs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Checkmate API
  version: 1.0.0
  summary: Self-hosted personal task manager.
  description: |
    A single-user task manager organised around four parallel **contexts**
    (Upsun, Personal, Gaal, Arkea): capture without choosing a context, triage
    deliberately, and read one daily brief.

    ## Authentication

    Three credential kinds, all resolving to the same user:

    | Kind | Used by | Obtained |
    |---|---|---|
    | Session cookie | Web | Google OIDC via `/auth/login/google` |
    | Device token (`cm_…`) | iOS, macOS, CLI | `POST /v1/tokens` (session only) or the CLI |
    | OAuth access token (`cmat_…`) | Remote MCP clients | `/oauth/authorize` + `/oauth/token` |

    Scopes are `read` (all GETs) and `write` (all mutations). A session cookie
    always carries both.

    ## Conventions

    - Collections return `{"data": [...], "next_cursor": "<id>"|null}`; single
      resources return the bare object.
    - Ids are UUIDv7 strings. Because v7 sorts chronologically, `id DESC` is
      creation order, which is how every list is ordered.
    - Timestamps are RFC3339 UTC (`2026-07-25T14:03:11.482Z`). Calendar dates are
      plain `YYYY-MM-DD` with no timezone.
    - Estimates are whole minutes.
    - Errors are `{"error": "...", "fields": {"due_on": "..."}}`. When `fields` is
      present, render each entry against its control rather than as a toast.

    ## Ordering

...
    opens with undated tasks.

...
  contact:
    name: Checkmate
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /v1/contexts:
    get:
      tags: [Contexts]
      summary: List contexts
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/includeDeleted"
        - name: include_archived
          in: query
          description: Archived contexts are excluded by default.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      tags: [Contexts]
      summary: Create a context
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContextCreate" }
      responses:
        "201":
          description: Created.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Context" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: |
            Validation failed. `slug` when the derived or supplied slug is already
            used by another live context.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A device token (`cm_…`) or an OAuth access token (`cmat_…`). An OAuth token
        is additionally checked to have been issued for this resource. A refresh
        token presented here is refused.
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-checkmate_session
      description: |
        Set by the sign-in callback. `HttpOnly`, `Secure`, `SameSite=Lax`. The
        `__Host-` prefix requires `Secure`, so a development server over plain
        http uses the unprefixed name `checkmate_session` instead.

        Cookie-authenticated mutations must additionally prove same origin, via
        `Sec-Fetch-Site` or a matching `Origin`; a request with neither is refused.
    oauth2:
      type: oauth2
      description: For remote MCP clients. PKCE with S256 is mandatory.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            read: See tasks, projects, contexts and people.
            write: Create, change and delete.
            offline_access: Receive a refresh token.
  parameters:
...
    limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from the previous page. **Opaque — do not parse it.**

        On a sorted task listing it is a keyset position bound to that exact sort,
        so reusing it with a different `sort` or `order` is a 422 rather than a
        quietly wrong page.
      schema: { type: string }
    includeDeleted:
      name: include_deleted
      in: query
      description: |
        Include tombstoned rows, which carry `deleted_at`. There is no undelete
        endpoint, so these are historical only.
      schema: { type: boolean, default: false }
  responses:
    BadRequest:
      description: |
        Malformed JSON, an unknown field, or more than one JSON document. Unknown
        fields are rejected rather than ignored, so a misspelled key on PATCH is
        reported instead of looking like a successful no-op.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        No credential, or one that is unknown, revoked, expired, or issued for a
        different resource. `WWW-Authenticate` names the protected resource
        metadata document and the scopes the request needs.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The credential lacks the scope this method needs — `read` for GET, `write`
        for a mutation — or a cookie-authenticated mutation failed the same-origin
        check. A scope failure carries `WWW-Authenticate` with
        `error="insufficient_scope"`, which is what lets an OAuth client step up.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationFailed:
      description: |
        Well-formed but not applicable. `fields` maps each offending input to a
        message; render them inline rather than as one toast.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            task:
              value:
                error: validation failed
                fields:
                  due_on: must be a YYYY-MM-DD date
                  title: is required
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: RFC3339, UTC, milliseconds.
      examples: ["2026-07-25T14:03:11.482Z"]
    Rev:
      type: integer
      format: int64
      description: |
        A global change counter, unique across every table and monotonic. Used as
        the sync cursor. Not writable.
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        fields:
          type: object
          description: Per-input messages, keyed by the request field name.
          additionalProperties: { type: string }
...
    Context:
      type: object
      required: [id, name, slug, sort_order, created_at, updated_at, rev]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        color:
          type: ["string", "null"]
          pattern: '^#[0-9a-fA-F]{6}$'
          description: |
            Exactly `#rrggbb`. Three-digit shorthand and eight-digit alpha are
            rejected rather than normalised, so a client never has to handle three
            spellings of one colour. Null means "use the fallback".

            Validated at the API edge rather than by a CHECK constraint: adding one
            would mean rebuilding the contexts table, and with foreign keys on that
            fires ON DELETE CASCADE into projects, recurrences and tasks.
          examples: ["#6b46c1"]
        sort_order:
          type: integer
          format: int64
          description: |
            Display order, writable. Contexts are the only entity with a
            persistable order; tasks and projects have none.
        archived_at:
          oneOf: [{ $ref: "#/components/schemas/Timestamp" }, { type: "null" }]
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
        deleted_at:
          oneOf: [{ $ref: "#/components/schemas/Timestamp" }, { type: "null" }]
        rev: { $ref: "#/components/schemas/Rev" }
    ContextCreate:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1 }
        slug:
          type: string
          description: Derived from the name when omitted. Unique per user.
        color:
          type: ["string", "null"]
          pattern: '^#[0-9a-fA-F]{6}$'
          examples: ["#6b46c1"]
        sort_order:
          type: ["integer", "null"]
          format: int64
          description: Appended after the existing contexts when omitted.
...
    ContextList:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Context" }
        next_cursor: { type: ["string", "null"] }
...